ref: move storage and email responses into their own directories
