+++ frontend/src/features/admin-form/responses/ResponsesPage/common/EmptyResponses/EmptyResponsesSvgr.tsx
