+++ frontend/src/features/admin-form/responses/ResponsesPage/common/ResponsesTabWrapper.tsx
