+++ frontend/src/features/admin-form/responses/ResponsesPage/storage/SecretKeyVerification.tsx
@@ -21,7 +21,7 @@ import Link from '~components/Link'
 
 import { FormActivationSvg } from '~features/admin-form/settings/components/FormActivationSvg'
 
-import { useResponsesContext } from './ResponsesContext'
+import { useResponsesContext } from '../ResponsesContext'
 
 const SECRET_KEY_NAME = 'secretKey'
 const SECRET_KEY_REGEX = /^[a-zA-Z0-9/+]+={0,2}$/
